feat(line,path): 응답/요청 dto에 추가된 필드 추가 및 관련부분 수정
- LineRequest : 추가요금 필드 추가
- PathResponse : 경로 거리, 요금 필드 추가
- 추가요금의 경우 nullable=true 이므로 null 입력가능하도록
  ExtraFare 의 변수 자료형 Double로 수정

+++ src/main/java/nextstep/subway/line/domain/ExtraFare.java
@@ -8,10 +8,10 @@ import javax.persistence.Embeddable;
 public class ExtraFare {
 
     @Column
-    private double extraFare;
+    private Double extraFare;
 
     public ExtraFare() {
-        extraFare = 0;
+        extraFare = (double) 0;
     }
 
     public ExtraFare(double extraFare) {
